+++ japps/Configuration.py
@@ -1,16 +1,20 @@
-from typing import Type
+from __future__ import annotations
+from os import PathLike
 import logging
+from pathlib import Path
 
-from typing import List, Dict
+from typing import List, Dict, Type, TYPE_CHECKING
+import japps.Errors
 
-from japps.plugins.IPlugin import IPlugin
-from japps.plugin_configs.IParser import IParser
-from japps.runners.IRunner import IRunner
 
+if TYPE_CHECKING:
+    from japps.plugin_configs.IParser import IParser
+    from japps.plugins.IPlugin import IPlugin
+    from japps.runners.IRunner import IRunner
 
 class Configuration:
-    plugins_directory: str = "plugins"
-    site_packages_directory: str = "site-packages"
+    plugins_directory: PathLike = Path("plugins")
+    site_packages_directory: PathLike = Path("site-packages")
 
     ignore_patterns: List[str] = ["__pycache__"]
     package_plugin_info_filename: str = "config.json"