+++ japps/Errors.py
@@ -12,3 +12,23 @@ class PluginNameError(PluginError):
 
 class CallTypeError(PluginError):
     ...
+
+
+class PluginImportError(PluginError):
+    ...
+
+
+class PluginConfigError(PluginError):
+    ...
+
+
+class PluginNoInfoError(PluginConfigError):
+    ...
+
+
+class RunnerError(Exception):
+    ...
+
+
+class UnknownPluginTypeError(RunnerError):
+    ...