+++ japps/PluginLoaders.py
@@ -1,11 +1,16 @@
 from abc import ABCMeta, abstractmethod
 
 from os import PathLike
+from os.path import relpath
+
+from typing import Dict
 
 from japps.Configuration import Configuration
 from japps.plugins.IPlugin import IPlugin
 from japps.plugin_configs.OneFile import OneFileParser
 from japps.Utils import import_from_path, install_dependencies
+from japps.Log import log
+from japps.Errors import PluginConfigError
 
 
 class IPluginLoader(metaclass=ABCMeta):
@@ -38,14 +43,25 @@ class SinglePluginLoader(IPluginLoader):
 
 class PackagePluginLoader(IPluginLoader):
     def parse_plugin(self) -> IPlugin:
+        log.debug(f"Loading plugin from package {self.path}")
         tmp_plugin = self.config.plugin_class()
 
-        default_data: dict = self.config.no_info_default.copy()
-        parsed_data: dict = self.config.package_plugin_info_parser.parse(self.path, self.config)
+        default_data: Dict[str, str] = self.config.no_info_default.copy()
+        parsed_data: dict = {}
+        try:
+            parsed_data = self.config.package_plugin_info_parser.parse(self.path, self.config)
+        except FileNotFoundError:
+            log.debug(f"Can't get config from package \"{relpath(self.path)}\", using config default")
+
         for key in default_data.keys():
             if key in parsed_data.keys():
                 continue
             parsed_data[key] = default_data[key].format(num=hash(self.path))
+
+        if not (parsed_data.get("NAME")
+                or parsed_data.get("TYPE")):
+            raise PluginConfigError(f"Can't find name or type in config from package {relpath(self.path)}")
+
         for key, value in parsed_data.items():
             setattr(tmp_plugin, key, value)