+++ japps/plugin_configs/IParser.py
@@ -1,9 +1,14 @@
-from abc import ABC, abstractstaticmethod
+from __future__ import annotations
+from abc import ABC, abstractmethod
+from os import PathLike
 
-from typing import Dict
+from typing import Dict, TYPE_CHECKING
 
+if TYPE_CHECKING:
+    from japps.Configuration import Configuration
 
 class IParser(ABC):
-    @abstractstaticmethod
-    def parse(path, config) -> Dict:
+    @staticmethod
+    @abstractmethod
+    def parse(path: PathLike, config: Configuration) -> Dict:
         ...