+++ japps/plugin_configs/OneFile.py
@@ -1,9 +1,12 @@
 import ast
 
-from typing import Dict, List
+from typing import Dict, List, Union
 
 from japps.plugin_configs.IParser import IParser
 
+
+def _filter_key(x: ast.AST) -> bool:
+    return isinstance(x, ast.Assign)
 
 class OneFileParser(IParser):
     @staticmethod
@@ -11,15 +14,16 @@ class OneFileParser(IParser):
         plugin_file = open(path)
         plugin_content = plugin_file.read()
         dump = ast.parse(plugin_content)
-        lines: List[ast.Assign] = list(filter(lambda x: isinstance(x, ast.Assign), ast.iter_child_nodes(dump)))
+        lines: List[ast.Assign] = list(filter(_filter_key, ast.iter_child_nodes(dump)))  # type: ignore[arg-type]
 
         result = {}
+        values: Union[List, str]
         for line in lines:
             if isinstance(line.value, ast.List):
-                values = [const.value for const in line.value.elts]
+                values = [const.value for const in line.value.elts]  # type: ignore[attr-defined]  # no it does
             elif isinstance(line.value, ast.Name):
                 values = line.value.id
             else:
-                values = line.value.value
-            result[line.targets[0].id] = values
+                values = line.value.value  # type: ignore[attr-defined]  # no it does
+            result[line.targets[0].id] = values  # type: ignore[attr-defined]  # no it does
         return result