+++ japps/plugin_configs/Py.py
@@ -1,4 +1,5 @@
-from typing import Dict
+from pathlib import Path
+from typing import Dict, Any
 
 from japps.plugin_configs.IParser import IParser
 from japps.Utils import import_from_path
@@ -6,8 +7,9 @@ from japps.Utils import import_from_path
 
 class PyParser(IParser):
     @staticmethod
-    def parse(path, config) -> Dict:
-        plugin = import_from_path(path + "\\" + config.package_plugin_info_filename)
+    def parse(path, config) -> Dict[str, Any]:
+        path = Path(path)
+        plugin = import_from_path(path / config.package_plugin_info_filename)
         return {
             key: getattr(plugin, key)
             for key in dir(plugin)